First commit some objects are done
A very simple model is adopted. The goal is to represent
graphs as close as possible to mathematical notation.
Once a graph is created it is immutable. So no adding
of nodes, edges etc.
However the json parser should be adopted so that
the vertices who have nil for edge ids are added
only to vertices and not even considered for edges.
It should already be the case but I should check it
anyway.

Here are several list of functions that should be added
for the next commit:

- filter vertex by id in graph
- filter edge by id in graph
- Get vertices by ids
- Get edges by ids
- make sure edge ids and vertice ids are unique.
- get avergare number of edges per vertex
- get max number of edges per vertex
- get min number of edges per vertex.
- check path for given two vertices
- Get distance between two vertices.
- Check cycles in graph
etc etc