VertexData type added. GraphML implementation is on going
VertexData type is added instead of using JsonNode.
VertexData mimicks json node with less functionality.
A subset of graphml would be used for graph io.

Maker needs to be updated accordingly. Its tests need
rewriting as well.

New test tasks are added to nimble file for facilitating
launching of single test files.
Vertex now uses VertexData as its data type. Maker needs
to see the type of data in graphML file stored as keys and choose
the correct data type for initalizing VertexData.

Test json files needs to be changed to graphml files.
The ids are changed to string ids in order to be coherent
with graphml ids