Graph objects are more or less done.
Graph objects are now more or less done. They mimick closely
mathematical objects.
Vertices of a graph is now a set of vertex.
Edges of a graph is now a set of edges. That is no ordering on
elements.
Containment is based on equality.
All objects are hashable, all objects now have an equality check
All objects except maybe graph itself can be represented as a
string.

The next big thing is to implement a graphml interface for
making graph io.
Not complex, but mostly very simple stuff.