New structure hyperedge added graph tests advanced
New structure hyperedge is added. It is a generalization
of edge. It supports simple methods. Its tests are to
be written.
Several examples in graphml are given for hypergraphs
and for not completed graphs.
isCompleted graph function returns true now I should
continue with the analysis.nim
Maybe add some methods for maker for reading hypergraphs
but that's about it.